Use decimal for customer monetary balance

+++ POS.Domain/Models/Customer.cs
@@ -114,8 +114,8 @@
             }
         }
 
-        private double _previousBalance;
-        public double PreviousBalance
+        private decimal _previousBalance;
+        public decimal PreviousBalance
         {
             get => _previousBalance;
             set
